requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.4
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
pytest-yaml==1.2.1
pywin32==228
PyYAML==5.3.1
selenium==3.141.0
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.3
urllib3==1.25.10
zipp==3.3.0
